DSND-2874: Move consumer test classes into kafka package
