+++ src/test/java/uk/gov/companieshouse/exemptions/delta/kafka/ConsumerInvalidTopicTest.java
@@ -1,9 +1,6 @@
-package uk.gov.companieshouse.exemptions.delta.kafka.consumer;
+package uk.gov.companieshouse.exemptions.delta.kafka;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
 import static uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils.COMPANY_EXEMPTIONS_DELTA_ERROR_TOPIC;
 import static uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils.COMPANY_EXEMPTIONS_DELTA_INVALID_TOPIC;
 import static uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils.COMPANY_EXEMPTIONS_DELTA_RETRY_TOPIC;
@@ -11,9 +8,8 @@ import static uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils.COMPANY_EX
 
 import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.time.Duration;
-import java.util.concurrent.TimeUnit;
+import java.util.concurrent.Future;
 import org.apache.avro.io.DatumWriter;
 import org.apache.avro.io.Encoder;
 import org.apache.avro.io.EncoderFactory;
@@ -22,6 +18,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.clients.producer.RecordMetadata;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -30,26 +27,19 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.kafka.test.utils.KafkaTestUtils;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import uk.gov.companieshouse.delta.ChsDelta;
-import uk.gov.companieshouse.exemptions.delta.exception.NonRetryableException;
-import uk.gov.companieshouse.exemptions.delta.kafka.AbstractKafkaTest;
-import uk.gov.companieshouse.exemptions.delta.service.ServiceRouter;
-import uk.gov.companieshouse.exemptions.delta.kafka.KafkaUtils;
 
 @SpringBootTest
 @WireMockTest(httpPort = 8888)
-class ConsumerNonRetryableExceptionTest extends AbstractKafkaTest {
+class ConsumerInvalidTopicTest extends AbstractKafkaTest {
 
     @Autowired
     private KafkaConsumer<String, byte[]> testConsumer;
+
     @Autowired
     private KafkaProducer<String, byte[]> testProducer;
+
     @Autowired
     private ConsumerAspect consumerAspect;
-
-    @MockitoBean
-    private ServiceRouter router;
 
     @DynamicPropertySource
     static void props(DynamicPropertyRegistry registry) {
@@ -64,19 +54,16 @@ class ConsumerNonRetryableExceptionTest extends AbstractKafkaTest {
     }
 
     @Test
-    void testRepublishToInvalidMessageTopicIfNonRetryableExceptionThrown() throws InterruptedException, IOException {
+    void testPublishToInvalidMessageTopicIfInvalidDataDeserialised() throws Exception {
         //given
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         Encoder encoder = EncoderFactory.get().directBinaryEncoder(outputStream, null);
-        DatumWriter<ChsDelta> writer = new ReflectDatumWriter<>(ChsDelta.class);
-        writer.write(new ChsDelta("{}", 0, "context_id", false), encoder);
-        doThrow(NonRetryableException.class).when(router).route(any());
+        DatumWriter<String> writer = new ReflectDatumWriter<>(String.class);
+        writer.write("hello", encoder);
 
         //when
-        testProducer.send(new ProducerRecord<>(COMPANY_EXEMPTIONS_DELTA_TOPIC, 0, System.currentTimeMillis(), "key", outputStream.toByteArray()));
-        if (!consumerAspect.getLatch().await(30L, TimeUnit.SECONDS)) {
-            fail("Timed out waiting for latch");
-        }
+        Future<RecordMetadata> future = testProducer.send(new ProducerRecord<>(COMPANY_EXEMPTIONS_DELTA_TOPIC, 0, System.currentTimeMillis(), "key", outputStream.toByteArray()));
+        future.get();
 
         //then
         ConsumerRecords<?, ?> consumerRecords = KafkaTestUtils.getRecords(testConsumer, Duration.ofSeconds(10L), 2);
